requirements.txt
opencv-python
numpy
tifffile
matplotlib
seaborn
Pillow
reportlab
tqdm
scipy